chore(map-service): group명 변경
- 그룹명을 on.logistics로 변경
