+++ map-service/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.7'
 }
 
-group = 'org.O-n-Logistics'
+group = 'on.logistics'
 version = '0.0.1-SNAPSHOT'
 
 java {
