+++ map-service/src/main/java/on/logistics/mapservice/MapServiceApplication.java
@@ -1,0 +1,13 @@
+package on.logistics.mapservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MapServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MapServiceApplication.class, args);
+    }
+
+}
