+++ map-service/src/main/java/on/logistics/mapservice/application/service/DirectionsService.java
@@ -1,0 +1,8 @@
+package on.logistics.mapservice.application.service;
+
+import on.logistics.mapservice.presentation.dtos.DirectionsResponseDto;
+
+public interface DirectionsService {
+
+    DirectionsResponseDto getRoute(String start, String end, String option, int cartype);
+}
