+++ map-service/src/main/java/on/logistics/mapservice/application/service/GeocodingService.java
@@ -1,0 +1,8 @@
+package on.logistics.mapservice.application.service;
+
+import on.logistics.mapservice.presentation.dtos.GeocodingResponseDto;
+
+public interface GeocodingService {
+
+    GeocodingResponseDto getCoordinates(String query);
+}
