+++ map-service/src/main/java/on/logistics/mapservice/global/config/WebClientConfig.java
@@ -1,0 +1,15 @@
+package on.logistics.mapservice.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Configuration
+public class WebClientConfig {
+
+    @Bean
+    public WebClient webClient() {
+        return WebClient.builder().build();
+    }
+
+}
