+++ map-service/src/main/java/on/logistics/mapservice/global/exception/ExceptionCode.java
@@ -1,0 +1,12 @@
+package on.logistics.mapservice.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface ExceptionCode {
+
+    String name();
+
+    HttpStatus getHttpStatus();
+
+    String getMessage();
+}
