+++ map-service/src/main/java/on/logistics/mapservice/presentation/dtos/GeocodingResponseDto.java
@@ -1,0 +1,13 @@
+package on.logistics.mapservice.presentation.dtos;
+
+import lombok.Builder;
+
+@Builder
+public record GeocodingResponseDto(
+    String roadAddress,
+    String jibunAddress,
+    String longitude,
+    String latitude
+) {
+
+}
