--- map-service/src/main/java/org/onlogistics/mapservice/MapServiceApplication.java
@@ -1,13 +1,0 @@
-package org.onlogistics.mapservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MapServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MapServiceApplication.class, args);
-    }
-
-}
