--- map-service/src/main/java/org/onlogistics/mapservice/application/service/DirectionsService.java
@@ -1,8 +1,0 @@
-package org.onlogistics.mapservice.application.service;
-
-import org.onlogistics.mapservice.presentation.dtos.DirectionsResponseDto;
-
-public interface DirectionsService {
-
-    DirectionsResponseDto getRoute(String start, String end, String option, int cartype);
-}
