--- map-service/src/main/java/org/onlogistics/mapservice/application/service/GeocodingService.java
@@ -1,8 +1,0 @@
-package org.onlogistics.mapservice.application.service;
-
-import org.onlogistics.mapservice.presentation.dtos.GeocodingResponseDto;
-
-public interface GeocodingService {
-
-    GeocodingResponseDto getCoordinates(String query);
-}
