--- map-service/src/main/java/org/onlogistics/mapservice/global/config/WebClientConfig.java
@@ -1,15 +1,0 @@
-package org.onlogistics.mapservice.global.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    public WebClient webClient() {
-        return WebClient.builder().build();
-    }
-
-}
