--- map-service/src/main/java/org/onlogistics/mapservice/global/exception/CustomException.java
@@ -1,14 +1,0 @@
-package org.onlogistics.mapservice.global.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-    private final ExceptionCode exception;
-
-    public CustomException(ExceptionCode e) {
-        super(e.getMessage());
-        this.exception = e;
-    }
-}
