--- map-service/src/main/java/org/onlogistics/mapservice/global/exception/ExceptionCode.java
@@ -1,12 +1,0 @@
-package org.onlogistics.mapservice.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public interface ExceptionCode {
-
-    String name();
-
-    HttpStatus getHttpStatus();
-
-    String getMessage();
-}
