--- map-service/src/main/java/org/onlogistics/mapservice/presentation/dtos/GeocodingResponseDto.java
@@ -1,13 +1,0 @@
-package org.onlogistics.mapservice.presentation.dtos;
-
-import lombok.Builder;
-
-@Builder
-public record GeocodingResponseDto(
-    String roadAddress,
-    String jibunAddress,
-    String longitude,
-    String latitude
-) {
-
-}
