+++ map-service/src/test/java/on/logistics/mapservice/MapServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package on.logistics.mapservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MapServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
