--- map-service/src/test/java/org/onlogistics/mapservice/MapServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package org.onlogistics.mapservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MapServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
